fix nodes overflow

+++ src/core.hpp
@@ -17,7 +17,7 @@
 class Engine {
     public:
 
-    int nodes = 0;
+    int64_t nodes = 0;
     int current_depth = 0;
     std::atomic<SearchLimit> limit;
     std::atomic<bool> interrupt_flag = false;
